# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Education Center CRM API
  version: 1.0.0
  description: REST API for the MVP CRM platform of an education center.
servers:
  - url: http://localhost:4000/api
...
paths:
...
  /lessons:
    get:
      tags: [Lessons]
      summary: List lessons with filters
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: date
          schema:
            type: string
            format: date
        - in: query
          name: age
          schema:
            type: integer
        - in: query
          name: title
          schema:
            type: string
      responses:
        "200":
          description: List of lessons
    post:
      tags: [Lessons]
      summary: Create a lesson
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonPayload"
      responses:
        "201":
          description: Lesson created
...
  /payments/create-intent:
    post:
      tags: [Payments]
      summary: Create Stripe or mock payment intent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentPayload"
      responses:
        "201":
          description: Payment created
  /payments/my:
    get:
      tags: [Payments]
      summary: List current parent payments
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Parent payments
...
  /feedback:
    get:
      tags: [Feedback]
      summary: List feedback threads for current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Feedback thread list
    post:
      tags: [Feedback]
      summary: Create a feedback thread
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Feedback thread created
  /feedback/options:
    get:
      tags: [Feedback]
      summary: List recipients and children for feedback forms
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Feedback form options
  /feedback/{id}:
    get:
      tags: [Feedback]
      summary: Get feedback thread details
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Feedback thread
  /feedback/{id}/messages:
    post:
      tags: [Feedback]
      summary: Add a reply to a feedback thread
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Feedback reply created
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LessonPayload:
      type: object
      required:
        - title
        - ageMin
        - ageMax
        - date
        - startTime
        - endTime
        - capacity
        - teacherName
        - price
      properties:
        title:
          type: string
        description:
          type: string
        ageMin:
          type: integer
        ageMax:
          type: integer
        date:
          type: string
          format: date-time
        startTime:
          type: string
        endTime:
          type: string
        capacity:
          type: integer
        teacherName:
          type: string
        price:
          type: number
...
    CreatePaymentPayload:
      type: object
      required: [enrollmentId]
      properties:
        enrollmentId:
          type: string
        currency:
          type: string
...